Videos respect the UI's current filtering mode.
Videos use point sampling when the UI is in point sampling mode, and smooth sampling in bi/trilinear modes. Allows a more authentic experience for people using point-sampled UI.

This does eliminate the current shipping configuration of point-scaled UI and smooth videos, though. It might be better to expose a separate filtering setting for the videos?