Removed ExecuteInEditMode from DaggerfallUnity
Should be able to work without this, as long as editor drives a couple
of checks from OnInspectorGUI.
Required checks are just to make sure Daggerfall path is operational and
singleton is connected to source binary data.
May need to revisit this again after more testing, but should not affect
gameplay as editor classes not present in builds.