Small change to support higher resolution UI
Previously, UI system was locked to 320x200 in DaggerfallBaseWindow.
This can now be set at construction time to allow higher resolution
windows on a case-by-case basis. Classic UI would still need a full
repaint and new fonts to support this correctly, which is not on roadmap
at this time.
First window to use higher resolutions will be quest inspector for
visualizing and debugging active quests.