Implemented the potion recipe viewer
Given a potion recipe in the inventory window, the user can now click
INFO on it and read the recipe. It pops up a new
DaggerfallPotionRecipeWindow, which shows the recipe name. Clicking
again then pops up a child DaggerfallMessageBox that shows the list of
ingredients in that recipe. This emulates the Daggerfall behavior.

Additionally, clicking USE on the recipe in the inventory window causes
"You cannot use this." to be displayed, which is also observed in
Daggerfall.