Lets the caller override the global alpha for the dim color of any PopupWindow
Previously the global alpha that's set in
DaggerfallUnity.Settings.DimAlphaStrength was used no matter what the
user specified in the Color32.

Changing this was necessary to emulate the behavior of Daggerfall when
reading potion recipes, which display two popup windows in a row. The
second window was dimming the screen a second time, whereas it doesn't
do that in Daggerfall.

This code chnage now uses the global DimAlphaStrength as a default
value, but if a caller sets their own Color32 for screenDimColor, it'll
honor that value.

This has the potential of changing existing popup behavior, though I've
tested it out with a few features and it dimmed correctly.