Use weather table from Daggerfall Chronicles
Weather is loaded from a JSON table based on the table in Chronicles (pg.49).

Currently weather is determined randomly. This is a bit jarring, however. It's possible to go from snowing to sunny with no transition. Marking it as a TODO to handle smoother/believable weather transitions.

Also added fog to the possible weather states.