Quest system progress
Experimenting with conditions as just another kind of action so they
share a common class structure.
Stubbed out "pc at" condition.
Added more boilerplate to ActionTemplate class to minimise how much code
is needed for an empty action class.
Added SetComplete() and IsComplete to IQuestAction and ActionTemplate.
This allows all one-and-done actions to set themselves as closed in a
standard way.
Added a short delay to QuestMachine startup to allow for future and to
give HUD fade a bit of time to work.
Added "say" action popup text.
Added LastStartMethod to StartGameBehaviour for future checks.
The two bootstrap quests _BRISIEN and _TUTOR__ are now started along
with every new character.