Added random NPC names to info clicks
The correct seed value for RMB and RDB random NPCs is currently unknown.
Using their offset position in game files to ensure a consistent result
for each NPC.
This can be changed to correct seed value once discovered. It probably
doesn't matter if names of random NPCs don't exactly match classic, but
it is a preference if possible.