RaceTemplate additions to support vamp/were races
Added flag for compound race. These are races which also have a base
race on player entity (i.e. the birth race before becoming a vampire).
This flag will be used by paper doll layout classes to correctly
assemble player appearance.
Added new paper doll background for werecreature players when
transformed.
Vampire and werecreature races not enabled currently as support needs to
be added to paper doll and body morphology setup.
This is not a priority item at this time.