Initial game support for global variables
Added persistent global variables manager to PlayerEntity. Chose to use
a dictionary of structs here rather than a simple array of bools. This
is so more detail is available in saves and to debuggers.
Globalvars are saved/loaded with games. Older saves without this data
are just set to defaults on load.
Fixed Clock bug. Will no longer create a range if second value is less
than first.
Parser now handles global references with or without actions in body.
Established globalvars links in Task for set/unset. Globalvars are how
information is communicated between quests.