Progressed NPC quest handling
QuestNPCClickHandler is now added to all fixed individual NPCs. This is
because they can be involved in one or more quests even in their home
location. This differs from other quest NPCs that are singletons
generated for one quest only.
Place will no longer allow an atHome NPC to be assigned to a site.
When clicking on an individual NPC, QuestNPCClickHandler will notify all
quests tracking this named NPC.
Person can now create a faction alliance NPC.
Some restructure to NPC setup and name assignment in Person resource.
Replaced TaskReturn with QuestBreak. Sometimes its necessary to break
from entire quest execution to say something.
Quest.ShowMessagePopup() will break quest to show message immediately.
This is the closest to classic behaviour with the current UI and quest
machine tick design. May need to revisit this at a later date if it
causes quest flow issues.
Added helpers to QuestMachine to check if a factionID is an individual
NPC and if an individual NPC is reserved to any quest sites.
RDBLayout will again remove NPCs reserved to a quest site. This is much
closer to final design.
Updated __DEMO07 quest to show branching conditions in addition to NPC
placement.