Implemented start of town navigation data
This will be used both to find valid spawn positions for mobiles inside
cities and to help wandering NPCs move around obstacles like buildings,
trees, pools of water, etc.
This is considered a partial implementation for now, as I only need to
find a valid spawn position in cities. Will build this out more when
wandering NPCs are implemented.