Fixes for recurring triggers
The "daily from" and "when task" trigger conditions need to be always
on, even if already triggered. Added a new flag to make a permanently
recurring trigger condition that will turn owning task on and off based
on trigger evaluation.
This actually seems to be Daggerfall's default behaviour for triggers,
but this can cause other issues (such as repeatedly showing same message
when clicking an NPC). Daggerfall appears to compensate for this problem
by making say/prompt messages only able to fire one per quest.
My solution will allow messages to fire multiple times if needed and for
triggers to be selective in whether they should be a single-fire (with
reset available) or always on.
Will continue to monitor this for good emulation behaviour as quest
system progresses and rework if needed.