Implemented hot-placement of resources
Quest resources (foes/items/npcs) are usually placed when player
transitions into building or dungeon interior. The placement code is
triggered by PlayerEnterExit.
This had the effect of resources not being placed when player already
inside the target location when resource was supposed to be placed.
Moved checks from PcAt to Place and added IsPlayerHere() check.
Now when a resource is assigned to a Place via AssignQuestResource(),
IsPlayerHere() will be called and resource instantiated directly if
player already at Place.