Improve parsing of global variable link task body
Global variable links may or may not have a body.
Previously was assumed a line break would indicate body or not, but this
is not always the case (see Sx999).
Parser now reads ahead to look for another task header or line break to
determine if it should read a body following global link or not.
This prevents parser from treating subsequent lines as a task body when
no line break present.