Climate namebank
The namebank used per subclimate is not always the same as the visible
race of people wandering the streets. Sentinel and Tigonus for example
have Breton wandering NPC sprites with Redguard names.
This seems limited to Subtropical and Swamp subclimates for now. Other
climates appear to have race/namebank parity. At least now settings are
available to be refined later.