Fix for line wrap in single line TextLabel
Added WrapText property to TextLabel which defaults to false.
CreateLabelTexture() now calls CreateLabelTextureSingleLine() (old
behaviour) or CreateLabelTextureWrapped() (new behaviour with wrapping).
Set MultiFormatTextLabel.AddTextLabel() to enable wrapping, as this
seemed to always be OK with the multi-format text. Things only broke for
TextLabels intended to be a single line with cutoff width.
Allofich, this should result in a fix for save window and still result
in your popups having wrapped behaviour. Please let me know if any
issues.