Fix for angled door in dungeons
This piece does not carry usual "DOR" or "DDR" tag, instead carries
"NEW" tag. Not sure if this should be globally considered a door piece
at this time or not.
For now, just treating this piece as a door by model ID alone. Will
consider expanding to all "NEW" tags if this proves to consistently be a
door tag.