Add post processing stack 1.0.4
There has been some request for post-processing effects by modders.
These are difficult to add to a standalone mod without some of the base
functionality already available in core.
These features are not used by default, but can be made available as an
option.
As this is a standard Unity addon, integrating into base project for
others to use in future.