Store map-level BuildingSummary in PlayerEnterExit
Only valid when player is inside a building (shop, guild, residence,
etc.).
This allows other systems to get correct unique map-level building data
for current interior even without StreamingWorld present to service
BuildingDirectory queries.
BuildingSummary is passed to PlayerEnterExit when player enters a
building and is saved/loaded with games.
This means older saves taken inside a building will not carry this data.
Player will need to exit/enter again for data to be populated. Anyone
consuming PlayerEnterExit.BuildingSummary should try to handle this
eventuality (if possible). Will not be a problem for saves taken
outside, in dungeons, or new saves.