Set ListBox global colours to ListItem at create
Nystul - this assigns the global colours set on ListBox at time of "new
ListItem(textLabel)". This ensures that other windows using the ListBox
that were created before this change have their colour settings used.
It's still possible to override the ListItem colours later, and does not
seem to impact the talk window.
Please let me know if any issues with this change.