PlayerEnterExit now caches building discovery data
PlayerEnterExit was previously caching BuildingSummary when entering a
building. However this did not carry display name of building. Changed
this to use PlayerGPS.DiscoveredBuilding as it carries full name in
addition to quality, faction, etc.
PlayerEnterExit.BuildingDiscoveryData is considered undefined when
player not inside a building (e.g. outside or inside a dungeon).
Old saves will not carry this information until player exits and enters
building again.