Change enemy attack move speed to 1/3
Even as experienced player, enemies can be a bit too fast and relentless
at 1/2, especially when facing multiple foes. Changed this to 1/3 for
now and will wait for feedback. Might be necessary to drop to 1/4 again
later.
I still like this change however, and agree it helps make enemies more
interesting. Perhaps this can be tuned to the reaction time selected
during character creation in future?