Expand CreateFoe to support "send" action
This is almost identical except foe will only spawn when outside within
a town rect. Example of use is the zombie sent by King of Worms with a
letter for player.
Because these spawns have a very long timeframe, allow the pending spawn
to be created anytime like "create foe" but only allow placement when
player outside in towns.
This can make foe trigger after fast travel, which seems to match
observed begaviour in classic.