Handle item carry followed by quest reward offer
Sometimes a quest will ask player to carry an item and then return that
same item back to player as a reward. This requires item be unequipped
and removed from inventory so player can pick up again from quest giver
reward process.
Added UnequipItem() overload to remove a specic item no matter what slot
it's equipped to.
GivePc will now unequip item (if equipped), remove from inventory (if
held), and convert to a permanent item before offer.
This matches behaviour in Sx010 which never explicity makes item
permanent in quest script.
It makes sense for any item being offered as a reward to be cleared of
quest state beforehand, or it will be removed automatically when quest
is tombstoned.