Progess main quest stage 5
Add DetailsMacro support for Item resource. Appears to be the same as
name macro.
Add support for a real Place target as a home location for NPC. This is
for Sx009 where player must visit an NPC at their home, but no actual
Place resource is defined, and "place npc" is never called.
The only hint appears to be these NPCs are part of a specific Local_3.x
faction group at time of Person creation. These NPCs seem to be
automatically instantiated in world at the specific location.
Understanding is still being developed, this may change over time. Goal
right now is just to support back-end for Sx009.
Still required to place this automatic NPC at home location for purposes
of quest completion.