Auto-assign special NPCs to home
This is equivalent to just using a Place resource properly in the first
place, the entire workaround here is just to support wonky quest script.
I believe even Daggerfall has issues with this, as the contact NPC
always displays as a "noble orc". This bug does not occur in Daggerfall
Unity.