Fix "when available" bootstrap
PlayerActivate was exiting after detecting a QuestResourceBehaviour
click. This is a hold-over from older methods of click handling.
Allow execution to run so click can be processed on StaticNPC in
addition to QuestResourceBehaviour. The StaticNPC click is now aware
enough not to open dialog when NPC has an active listener.