Disable body placement in Sx006
This is problematic right now in a few ways:
* Daggerfall Unity separates quest and item spawns correctly by marker,
so body will not spawn in same position as NPC. The fact this worked n
classic is probably a quirk of quest engine, as original script was
never intended to work this way. Change was made in a community patch.
Question is whether I want to reproduce behaviour just for this one
quest.
* Possible to pick up Castellian's body as it's just a normal quest
item.
* Hot-placement of quest item causes quest Person to be re-injected to
scene. This is a bug on my side.
For now just not hot-placing body but still hiding Castellian. Quest can
still operate as normal.