Always place foe freely inside buildings
Feel this gives a better result in most cases - especially when player
is supposed to be beset by multiple foes at once. It didn't make sense
for those foes to be stashed at random throughout the building interior.
Leaving old method commented out in case I want to revisit this and tune
later.