Implement DaggerfallActionDoorSpecial
This is a secret action door component found on door-like models that
are intended to be opened only by lever. Player should not be aware of
door nor able to open, bash, pick, inspect, or cast their way through.
Opted to make this a cut-down version of standard door so no need to
work around all the gameplay logic built into normal doors everywhere
else.
When laying out an RDB block, the DaggerfallActionDoorSpecial component
is added when object carries an OpenDoor or CloseDoor action but is not
a normal action door.
DaggerfallAction will look for a special door first then a normal door
when OpenDoor/CloseDoor actions are called.
This type of door is found in many dungeons. An example is the SE pillar
going into Lysandus' Tomb which player must open from a lever below the
main tomb area.
This should resolve other known cases, such as "Stronghold of the
Blades" quest, where secret doors are supposed to open when player
stands near target item.