Link quest item spawn to live scene marker
When quest Place is instantiated a unique markerID is assigned from
enumerated markers.
This is used to parent quest Item (only items at this time) to the
actual live marker in scene at layout time.
Parenting allows the quest item to move with a translation action record
if present.
The most notable scenario is totem in Daggerfall treasury.
The 'tele2qitem' cheat will continue to teleport player to original
position for now. But the totem is currently only known case of mobile
quest item, so not a priority for now.
If no marker pairing is found, quest item is added using previous logic.