Add PlayerMotor.FreezeMotor property
This is a workaround to a physics latency issue. When player "fires"
themselves using Sheogorath's Crossbow, this is actually a teleport
followed by a translation of crossbow bolt. Player is supposed to end up
standing in notch after bolt fires.
However, player in Unity is able to fall back through moving platform
because it's moving so quickly. Added a universal motor freeze of 0.5
seconds and a movement cancel on teleport to clear any moving platform
attachments.
This allows player to safely land on bolt in final position as intended.
Will need to revisit this after motor refactor.