Track symbol of last Place a Person was assigned
When "place npc" is called, the associated Place symbol is now written
to the Person data. If the Person is relocated, then Place will be
updated.
Will need to restart quests to test. Previously instantiated quest data
from save games will not carry this information.
Added Person.GetAssignedPlaceSymbol() to retrieve the Place this Person
was last assigned to. Can be null if Person not yet assigned or
previously destroyed.
Added Person.IsPlayerInSameLocationWorldCell() as a quick way to check
if player in same world cell as Place this Person was assigned to. Does
not care about location type, building, or interior/exterior. Just
matches mapID. When player standing inside same world cell, mapID should
match.