Fix null exception when changing weapons
The fields WeaponManager.LeftHandWeapon and .RightHandWeapon are used
for rendering and animation only. They really mean "left-of-screen" or
"right-of-screen" for drawing purposes, nothing to do with which hand
player is actually holding a weapon in. This was originally added to
emulate dual-wielding back in the early DFTFU days by drawing two
weapons simultaneously. Daggerfall however only draws a single weapon
regardless of which hand it is actually held in.
Some of the newer weapon code is now referencing LeftHandWeapon which is
not used at all in the master scene. This means attempting to switch
weapons (default H) or issue attacks would cause exceptions and lock up
attacks.
I found this distinction a bit confusing between "weapons held" and
"weapon rendered" so decided to clean this up and simplify the concept.
Removed LeftHandWeapon and RightHandWeapon fields. Replaced with a
single ScreenWeapon field. This represents the weapon currently drawn on
the screen. ScreenWeapon will be set based on whichever weapon player is
actually using from the currently active hand.
Renamed FPSWeapon.LeftHand to .FlipHorizontal to better describe what's
happening.
Greatly simplified the WeaponManager.ExecuteAttacks() methods as it no
longer has to track dual-wield animations.
Merged IsLeftHandAttacking() and IsRightHandAttacking() to
IsWeaponAttacking() as only a single weapon is displayed now.