Wire up spell cast/recast handling
A spell selected from spellbook is sent to player's SpellManager as a
readySpell.
When a readySpell is pending, the weapon is deactivated and activate
clicks will fire spell.
When recast action is selected, the last spell cast will be assigned to
the readySpell.
No actual spells are being fired. This is just to get the front-end
interface and UI ready for spell casting.