Revert Foe resource changes
I feel this breaks the one-to-many design of Foe resource. This change
would have made it such that all "create foe" spawns inside quest had
same gender and name. There is no good way to treat Foe resource as both
one-to-many or individual without breaking one or the other. This is a
limitation of Daggerfall's quest design.
Quest authors can probably work around this manually, but would need the
ability to at least link gender of Person and Foe resources to ensure
spawned Foe matches dummy Person gender. This would be a Daggerfall
Unity specific enhancement.
Will return to this issue later.