Start work on effect bundle class
This class is like a container and execution context for effects grouped
together. A spell is an example of an effect bundle carrying one or more
effects.
Once an effect bundle attaches to an entity by way of effect manager, it
will be executed every tick of effect system until expired.