ModSettings: refactoring
Local values has been migrated from ini to json.
Similarly, presets are now serialized in a single json file, with the same
interface used for presets included inside mods, which is managed
automatically by in-game preset picker window. Additional presets
provided inside other files are imported as read only.

The ModSettingsData now includes all functionality of loading, saving and applying a preset,
reusing the same code for editor and game windows as much as possible. As a result the class
ModSettingsReader has been removed.

Switch-cases in editor and game windows are been replaced by overrides directly inside
Key derived classes to make full use of polymorphism and allow a smooth addition of new controls in the future.
The KeyType enum is preserved as it's required to choose a control on editor gui.