Improve Headbobber support
Changes the statement that assigns the mainCamera.transform.localposition so that it adds the difference between it's current position and new position to the current position.
This allows other code from other classes to modify the localPosition of the camera without the Headbobber dominating the position.
This commit is being made to incease compatibility with other code I'm working on, without changing the functionality of headbobber.