Fix game loop crash when unable to resolve god name
The new %god name macro handling was looking for previous Person referenced by dialog stream. This would cause an exception stream when questor dialog requires %god without a previous Person reference (null MCP).
Added support to look for current questor, so this will be correct for quest dialog when handing in to Person who is questor.
And changed fall-through to use a random god name. Better to fill in something here than allow game to crash. This is a fairly minor macro and even partial support is enough for now.