Fix bow range on melee
When changing from bow to melee the range property was updated in SetWeapon() but not SetMelee() so range did not change for hand-to-hand. This allowed player to kick and punch things at bow range.
Changed this to be updated in ApplyWeapon() instead so it will always set range after changing weapons.
Added a null check and range constants.