Set entity defaults when starting a new game or loading
This makes entity reset more automatic than before.
Added isParalyzed flag to entity for any class implementing that entity to consume. Requirements will be a little different depending on which entity is paralyzed.
The isParalyzed flag is not serialized intentionally. This flag is intended to be controlled by effect system (which is serialized).
Likely to move isLevitating flag into Entity class at some point.