Workarounds for several commoner quests instantly ending
There are two problems here. First is that some commoner quests request an automatic location such as House4-House5 that don't seem to exist.
The second problem is Clock resource does not compute travel time for automatic locations as not flag & 16. The flag commonly used for these timers appears to be 1.
Added a workaround hack to convert House4-House5 back to House2 and generate travel time when flag & 1 in Clock resource and timer otherwise 0.
Need to test this change back against other quests to ensure doesn't break anything else.