Start work on unified resistance handling
Exposing entity resistances in a unified manner for effect system.
The way resistances are handled by classic back-end is not well known, and rather buggy as well.
Designing this in a way that should be more robust and integrate with the current entity-effect system.
Yet to implement actual resistance handling in entity-effect setup. Just setting up data management right now.