Remove initial magic round from effect Start()
This was causing an execution order problem for some effects.
Initial magic round is now performed by EntityEffectManager shortly after effect is started and attached to an entity.
This is basically identical to previous design except now effect has a chance to setup in Start() before first magic round is triggered.