Start changes for crouching bug, see description
The old serializablePlayer code would restore the IsCrouching variable upon loading a game.  But ever since the smooth crouching update, the code does not  "snap" the player into place like it used to.

If the player is standing, and in the loaded game the player is crouching, The player will be standing when loaded, move at crouching speed and be unable to properly completely crouch.

If player is crouching and in the loaded game the player is standing, The player will be crouched and move at standing speed.  Also trying to crouch will sink the player into the ground.

The plan is to copy old code from before the first implementation of "croucher.cs" (now PlayerHeightChanger) and use it to create a new method that can be called when save game is loaded to snap the player into the correct height.