Refine player repositioning when loading pre floating-y save
Reverted DefaultTerrainSampler.Version. This was too agressive a method and not necessary in this case. It also didn't solve the problem for users running alternate samplers.
Added FloatingOrigin.floatingOriginVersion to create a new sampler-independent means of repositioning player. This field will be saved with all new save games.
When player loads an exterior save game, they will just be placed on ground in their current position. Chances are player is already on ground and will not notice any difference.
Interiors and dungeons can restore position as normal.
Next job is to fix loot piles and anchor in teleport effect when loading downlevel floating origin saves.