Add ability to get miscellaneous work in castles
The code for this is simpler than the code for getting quests in towns. NPCs don't mention people in castles when asked about work so their data needn't be stored for dialogue. A separate dictionary is used to determine whether or not the PC previously talked to a given NPC in the castle.