Stub Transfer Effect
Changing behaviour of stat transfer compared to classic so that magnitude of effect is properly drained from target and transferred to caster. In classic this effect works identically to Drain as far as I can tell.
Transfer Health/Fatigue will be added later and work just like classic, as these function as expected.