First pass at Transfer effect
This effect establishes a vampiric link where a stat from target entity is transferred into the player permanently until target dies or passes out of scope (e.g. leaves area).
Created MergeDirectStatMods() methods for effects like this to send stat/skill mods directly to another enity. In this case used to reflect the drain back to player as a buff.
Direct stat mods are cleared at the start of every magic round, so effects using them must re-apply in their own magic round tick.
This effect is an improvement over classic which only drained target and did not actually transfer attribute to player. The goal is to make Transfer effect actually work in line with its description while remaining balanced and interesting to use. Will continue to refine over time.
One way this can be exploited right now is if player casts transfer on an entity then runs past them. It's very easy to buff everything to 100 this way. I will add a distance check later so effect will expire automatically the moment caster is out of a certain short range of target. So vampiric link can be broken by distance.
This also exposes an issue with deserialization of effect bundles on enemy entities. Will need to correct this before issuing next builds.