Start work on "create npc" quest action
Stubbed CreateNpc action and register to QuestMachine. Currently not functional.
NPCs added using CreateNpc are instantiated to their home Place. This means the current handling of creating a home location must expand beyond simple flavour text and create a real Place terget for every virtual NPC.
Add supporting Place.ConfigureFromPlayerLocation() method to allow a home Place to be generated from player's current world position. This was not possible using just script source.
Created Place.GetBuildingName() as own method as this may be useful to be called later.
A Person resource will now create a home Place resource for every single NPC. If this is a fixed/questor NPC then current player location is used to generate Place, as where player clicked NPC is where NPC lives.
For other NPCs they will be assigned a building based on faction and a scope based on local/remote hints. The NPC home Place is then generated as before.
Moved "BLANK" string to QuestResource as this can be used anywhere.
This code is all currently untested. Will start testing with quests now and submit fixes as problems are found.