Combine travel time when multiple Place resources
Some quests send the player to several places in the course of a single quest. This could result in deadlines that were impossible or very nearly impossibe as GetTravelTimeInSeconds() was only calculating travel time to first resource.
This also causes problems when most distant place is not first place in list.
Modified behaviour of default GetTravelTimeInSeconds() so that ALL place resources are combined for purposes of travel time. This ensures player should have time to complete quest.
This can result in quest times being overly generous. Will look at tuning in future.