Scaling fix for rendering panels to RenderTexture
Unity scales DrawTexture() rendering to RenderTargets based on display resolution.
Using the scaling features of custom UI system to compensate so panel is scaled appropriately. This is a bit of a hack, need to find a better solution (if one is available).
Can now set filter mode of target texture.