Support for primary UI redirection
Removed EnableDefaultUserInterface property from DaggerfallUI, as interface can no longer be disabled. It can however be redirected to another target output.
Added CustomRenderTarget and CustomMousePosition to DaggerfallUI to redirect output and feed in custom mouse coordinates.
Improved automatic unpause in UserInterfaceManager to account for HUD being open or not.
Replaced TargetRect property with TargetSize in UserInterfaceRenderTarget.