Fix Betony travel map crosshair placement
This required a substantial manual adjustment for Y axis and fix to scaling calculation.
Added a constant for Betony index as it requires a good bit of manual tuning based on differences in scale relative to other maps.
Betony should work fully on travel map now.
Cybiades has similar problems, but only a single dungeon location out to sea. Not used for any gameplay. Will do another pass on this later.