Refactor and remake HeightChanger
Change HeightChanger so that it is cleaner and easier to understand.

also, height change of controller is additive which means rather than set the height = to a specific number, the current difference between the controller height and target height gets added to the controller.  This enables other height changing actions to modify the height, like going into outside pools of water.