Add Vitals Loss indicator system
Each of the player's vitals on the hud now have a bar that visually and actively shows how much is lost of each bar after it happens.

If the player loses health, after the health bar updates, it exposes the indicator bar that was hiding behind the health bar, then the quickly and smoothly shrinks and hides behind the health bar.