Better interaction between bars
Loss and gain indicator bars can both be displayed at the same time and healing first or healing after doesn't hide the bars anymore unless the gain fills the guage, it will then hide the loss.

changed the bars so that if more gain or more loss is added, and the bar is still smoothing, then it will pause for a shorter time before continuing to smooth to target size.  Will also reset progress in lerp to the target amount.