Implement Underwater Fog
Implements a smooth underwater fog,  if the player is partially underwater, the fog fades partially in, until the camera is under.  This eliminates the blinking sensation that happens when exiting/entering water.

Removed a toggle for global fog fix for unity 5.5 because we're not unity 5.5 anymore, and it was preventing underwater fog from working.