Initial SDF font support in DagUI
Adding support for SDF (signed distance field) fonts to internal UI system. This does not rely on any external tools for rendering, only to pre-generate the SDF font asset.
The goal is to fully preserve classic formatting so that layout codes for books, etc. will continue to work as expected while enabling smoother fonts at higher resolutions.
The compromise is that glyphs may appear stretched or squished to match equivalent Daggerfall glyph widths in-place. The SDF shader ensures the font continues to look smooth and use as much pixel density is available in that rect however.
Shipping this as disabled while still under development. Currently only edit fields and tooltips will see the outcome.
This breaks the experimental HQTooltips option and version text when enabled. This will be corrected soon.