Fixes Falling in water when climbing
Fixes Falling in water while climbing by adding a buffer that the player cannot fail climbing above water level.

Allows the player to look left and right a little bit while climbing, but doesn't reduce the shallowness of the angle the player needs to look to start climbing.

Uses a CapsuleCast to tell if the player should stop climbing and start going forward, and gets the direction of the ledge the player will be climbing onto.  Forces the player's horizontalmove direction towards the ledge directly.