Implement Slip and catch system
Designed to be like the classic daggerfall system.

Player now has a chance to climb successfully, a chance to slip, and a lower chance to regain hold and continue climbing.  All chances can be offset by having a high skill in climbing.

80% base chance of success for continuing to climb per skill check.  A player with 50 climbing skill would boost this to 90% and 75 would boost to 95%

20% base chance of regaining hold (at 37% of climbing skill check's wait time).  a player with 50 skill would boost this chance to 60%, 75 skill boosts it to 80%

once the player is climbing, they can swivel around and look in any direction as long as they hold the forward key and keep in mind when they reach the top they'll move in direction they are facing!