Implement airborne grasp and jump system
Player can now jump off walls and try to grasp another wall with a moderate skill check (about halfway between the difficulty of the regain hold skill check and the continue climbing skillcheck.) You can also jump from the ground to try and grasp a wall you are facing.

The horizontal movement tolerance to start climbing is increased while airborne and returned to normal when on the ground.