Fix HUD crash
This happened in GuildManager.GetGuildGroup() when checking factionData.children.Count > 0. For factions with no children, this value can be null.
Adding a null check to before checking count to short circuit prevents this exception and code continues on as if faction has 0 children, which is going to be the case under these conditions.