Add image panel to DaggerfallMessageBox
This just adds the image Panel and exposes it via a property. Will need to set Size and BackgroundTexture properties to display image. Aligned to centre-top by default, but this also be changed by ImagePanel property.
Not sure how best to size image. For paintings, it's probably best defined by the number of rows available. So height would be rows available * messagePanel.Label.Font.GlyphHeight and width would be the correct aspect width based on that height result.
@ajrb I hope this helps get you started with adding images to paintings. I'm happy to help as much as needed here.